requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# F1 Data APIs
fastf1>=3.0.0
ergast-py>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: Dashboard (uncomment if needed)
# streamlit>=1.28.0
# dash>=2.14.0